pyproject.toml
[project]
name = "network-security-project"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2026.1.4",
    "dagshub>=0.6.6",
    "dvc>=3.66.1",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]
